platform/external/webkit/LayoutTests/fast/events/pop-state-event-creation-expected.txt

PASS document.createEvent('PopStateEvent') instanceof window.PopStateEvent is true
PASS document.createEvent('PopStateEvent') instanceof window.Event is true
PASS document.createEvent('PopStateEvent').constructor === window.PopStateEvent is true
PASS successfullyParsed is true

TEST COMPLETE
